Conflicts:
	extensions/ida2015/src/main/java/eu/amidst/ida2015/PlateuHiddenVariableConceptDrift.java

+++ huginlink/src/main/java/eu/amidst/huginlink/inference/HuginInference.java
@@ -40,7 +40,7 @@ public class HuginInference implements InferenceAlgorithm, Serializable {
      */
     private void setVarEvidence(Variable n, double value) throws ExceptionHugin {
         if (n.isMultinomial()){
-            ((DiscreteNode)huginBN.getNodeByName(n.getName())).selectState((long)value);
+            ((DiscreteNode)huginBN.getNodeByName(n.getName())).selectState((int)value);
         }
         else if (n.isNormal()) {
             ((ContinuousChanceNode)huginBN.getNodeByName(n.getName())).enterValue(value);
